Auto detect run_list
